API: Move repository set tests to a separate module
For UI/CLI we have separate module test_repository_set.py for repository set tests.
Making API consistent with other interfaces.
Closes #3148
